Globe Patching started
- Imported Earth asset from Asset Store
- Added 7 Track Points
- Every track point has a child earth

TODO:
- Only show one globe with multiple simultaneous trackpoints
- There is an offset problem with the globe when the track points are
picked up at different location. For an example, picking up the first
track point in USA will map the earth correct in relation to that track
point. If you then pick up a track point in Europe, the rotation of the
augmented earth will no longer correlate to the rotation of the
physical globe.